Add Incremental View Maintenance using counting algorithm
DISTINCT and tuple duplicates in views are supported. When IMMV is
created, multiplicity of tpules is counted and stored in  "__ivm_count__"
column, which is a hidden column of IMMV.   The value in__ivm_count__
is updated when IMMV is maintained incrementally. A tuple in IMMV
can be removed if and only if the count becomes zero.

When SELECT is issued for IMMV with tuple duplicates, each tuple is
returned __ivm_count__ times. generate_series(1, __ivm_count__)
function is used to implement this.  If IMMV is defiend sith DISTINCT
clause, tuples in these are returned directly without using
generate_series.

This patch also allows self-join, simultaneous updates of more than
one base tables, and multiple updates of the same base table.
I first tried to support just self-join, but I found that
Note this can be a basis of Deferred-maintenance implementation
planned in future.

In this implementation, AFTER triggers are used to collecting
tuplestores containing transition table contents. When multiple tables
are changed, multiple AFTER triggers are invoked, then the final AFTER
trigger performs actual update of the matview. In addition, BEFORE
triggers are also used to handle global information for view
maintenance. To calculate view deltas, we need both pre-state and
post-state of base tables. Post-update states are available in AFTER
trigger, and pre-update states can be calculated by filtering inserted
tuples using cmin/xmin system columns, and append deleted tuples which
are contained in a old transition table.